Bundeled everyhting inside appropriate widgets (components)
