+++ src/admin/widgets/Cutsomers.tsx
@@ -5,9 +5,11 @@ import {
 } from '../hooks/customer-analytics';
 import { Button, Container, Text } from '@medusajs/ui';
 import { PieChart } from '../components/PieChart';
-import { ReturningCustomers } from '../components/KPI';
-import { BarChartTypes, TopCustomerGroupBySales } from '../components/Charts';
 import { BarChartSkeleton } from '../skeletons/BarChartSkeleton';
+import { OrderAnalyticsResponse } from '../hooks/order-analytics';
+import { SmallCardSkeleton } from '../skeletons/SmallCardSkeleton';
+import { BarChart } from '../components/BarChart';
+import { ProductAnalyticsResponse } from '../hooks/product-analytics';
 
 const today = new Date();
 const daysPrior30 = new Date(new Date().setDate(today.getDate() - 30));
@@ -53,6 +55,53 @@ const CustomerWidget = () => {
       </div>
     </>
   );
+};
+
+type KPIProps<T = OrderAnalyticsResponse> = {
+  data: T | undefined;
+  isLoading: boolean;
+};
+
+const ReturningCustomers: React.FC<
+  KPIProps<CustomerAnalyticsResponse> & {
+    specificTimeline?: string;
+  }
+> = ({ data, isLoading, specificTimeline }) => {
+  return (
+    <Container className="relative">
+      <div className="flex justify-between items-center">
+        <Text size="large">
+          Returning Customers{' '}
+          {specificTimeline && (
+            <span className="text-ui-fg-muted">({specificTimeline})</span>
+          )}
+        </Text>
+        <a href="/app/analytics?tab=customers">
+          <Button
+            variant="transparent"
+            className="text-ui-fg-muted text-xs lg:text-sm"
+          >
+            View more
+          </Button>
+        </a>
+      </div>
+      {isLoading ? (
+        <SmallCardSkeleton />
+      ) : (
+        <>
+          <Text size="xlarge" weight="plus">
+            {data?.returning_customers || 0}
+          </Text>
+        </>
+      )}
+    </Container>
+  );
+};
+
+type BarChartTypes<T = ProductAnalyticsResponse> = {
+  data: T | undefined;
+  isLoading: boolean;
+  specificTimeline?: string;
 };
 
 const OrderBreakdownPie: React.FC<BarChartTypes<CustomerAnalyticsResponse>> = ({
@@ -106,6 +155,57 @@ const OrderBreakdownPie: React.FC<BarChartTypes<CustomerAnalyticsResponse>> = ({
     </Container>
   );
 };
+
+const TopCustomerGroupBySales: React.FC<
+  BarChartTypes<CustomerAnalyticsResponse>
+> = ({ data, isLoading, specificTimeline }) => (
+  <Container className="min-h-[9.375rem] ">
+    <div className="flex justify-between">
+      <div>
+        <Text size="xlarge" weight="plus">
+          Top Customer Groups by Sales
+        </Text>
+        <Text size="xsmall" className="mb-8 text-ui-fg-muted">
+          Sales breakdown by customer group in the{' '}
+          {specificTimeline ?? 'selected period'}
+        </Text>
+      </div>
+
+      <a href="/app/analytics?tab=customers#:~:text=Top%20Customer%20Groups%20by%20Sales">
+        <Button
+          variant="transparent"
+          className="text-ui-fg-muted text-xs lg:text-sm"
+        >
+          View more
+        </Button>
+      </a>
+    </div>
+    {isLoading ? (
+      <BarChartSkeleton />
+    ) : data?.customer_group && data.customer_group.length > 0 ? (
+      <div className="max-w-80 mx-auto aspect-video">
+        <BarChart
+          data={data.customer_group}
+          xAxisDataKey="name"
+          lineColor="#82ca9d"
+          useStableColors={true}
+          colorKeyField="name"
+          yAxisDataKey="total"
+          yAxisTickFormatter={(value: number) =>
+            new Intl.NumberFormat('en-US', {
+              currency: data.currency_code || 'EUR',
+              maximumFractionDigits: 0,
+            }).format(value)
+          }
+        />
+      </div>
+    ) : (
+      <Text size="xsmall" className="text-ui-fg-muted text-center">
+        No data available for the selected period.
+      </Text>
+    )}
+  </Container>
+);
 
 export const config = defineWidgetConfig({
   zone: 'customer.list.before',
